aulas/modulo4/parte1(1)/src/modules/users/services/UpdateProfileService.ts
import { inject, injectable } from 'tsyringe';

import User from '@modules/users/infra/typeorm/entities/User';
import AppError from '@shared/errors/AppError';
import IUsersRepository from '../repositories/IUsersRepository';
import IHashProvider from '../providers/HashProvider/models/IHashProvider';

interface IRequest {
    user_id: string;
    name: string;
    email: string;
    password?: string;
    old_password?: string;
}
@injectable()
class UpdateProfileService {
    constructor(
        @inject('UsersRepository')
        private usersRepository: IUsersRepository,

        @inject('HashProvider')
        private hashProvider: IHashProvider,
    ) {}

    public async execute({
        email,
        name,
        old_password,
        password,
        user_id,
    }: IRequest): Promise<User> {
        const user = await this.usersRepository.findById(user_id);

        if (!user) {
            throw new AppError('User not found');
        }

        const userWithUpdateEmail = await this.usersRepository.findByEmail(
            email,
        );

        if (userWithUpdateEmail && userWithUpdateEmail.id !== user_id) {
            throw new AppError('Email already in use');
        }

        user.name = name;
        user.email = email;

        if (password && !old_password) {
            throw new AppError('You need to inform the old password');
        }

        if (password && old_password) {
            const checkPassword = await this.hashProvider.compareHash(
                old_password,
                user.password,
            );
            if (!checkPassword) {
                throw new AppError('Old password doesnt match');
            }
        }

        if (password) {
            user.password = await this.hashProvider.generateHash(password);
        }

        return this.usersRepository.save(user);
    }
}

export default UpdateProfileService;
